pyproject.toml
[project]
name = "handson-mcp-and-agent"
version = "0.1.0"
description = "Hands-on lab for MCP and Agent development"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "PyPDF2==3.0.1",
    "pypdf==5.5.0",
    "mcp[cli]>=1.21.0",
    "fastmcp>=2.13.1",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.60",
    "langchain-huggingface==0.2.0",
    "langchain-mcp-adapters==0.1.1",
    "langchain-openai==0.3.17",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.5",
    "langchain-ollama>=0.3.3",
    
]
